fix corner case regression introduced by the new transport

+++ libopflex/comms/active_connection.cpp
@@ -167,6 +167,11 @@ void on_active_connection(uv_connect_t *req, int status) {
     peer->ai_next = NULL;
 #endif
 
+    if (peer->unchoke()) {
+        retry_later(peer);
+        return;
+    }
+
     peer->unlink();
     peer->insert(internal::Peer::LoopData::ONLINE);
